Move/rename initTestMigration
