+++ apps/api/src/db/index.ts
@@ -18,13 +18,6 @@
  */
 
 import { drizzle } from 'drizzle-orm/node-postgres';
-import { migrate } from 'drizzle-orm/node-postgres/migrator';
-import { dirname } from 'path';
-import { fileURLToPath } from 'url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-const migrationsFolder = __dirname + '/../../drizzle';
 
 export type PostgresDb = ReturnType<typeof drizzle>;
 
@@ -34,16 +27,6 @@ export const connectToDb = (connectionString: string): PostgresDb => {
 		return db;
 	} catch (err) {
 		console.log('Error on Database startup');
-		console.log(err);
-		throw err;
-	}
-};
-
-export const initMigration = async (db: PostgresDb) => {
-	try {
-		await migrate(db, { migrationsFolder });
-	} catch (err) {
-		console.log('Error Migrating on Database startup');
 		console.log(err);
 		throw err;
 	}
