+++ apps/api/src/tests/db/index.ts
@@ -23,7 +23,7 @@ import { after, before, describe, it } from 'node:test';
 import { PostgreSqlContainer, StartedPostgreSqlContainer } from '@testcontainers/postgresql';
 import { eq } from 'drizzle-orm';
 
-import { connectToDb, initMigration, type PostgresDb } from '../../db/index.ts';
+import { connectToDb, type PostgresDb } from '../../db/index.ts';
 import { actions } from '../../db/schemas/actions.ts';
 import { agreements } from '../../db/schemas/agreements.ts';
 import { applicationContents } from '../../db/schemas/applicationContents.ts';
@@ -31,7 +31,7 @@ import { collaborators } from '../../db/schemas/collaborators.ts';
 import { files } from '../../db/schemas/files.ts';
 import { revisionRequests } from '../../db/schemas/revisionRequests.ts';
 
-import { PG_DATABASE, PG_PASSWORD, PG_USER } from '../testUtils.ts';
+import { initTestMigration, PG_DATABASE, PG_PASSWORD, PG_USER } from '../testUtils.ts';
 
 describe('Postgres Database', () => {
 	let db: PostgresDb;
@@ -50,7 +50,7 @@ describe('Postgres Database', () => {
 
 		db = connectToDb(connectionString);
 
-		await initMigration(db);
+		await initTestMigration(db);
 	});
 
 	describe('Connection', () => {
